Changed package name, group and artifact id for support gradle built-in mapping plugin id to maven coordinates
