--- src/integrationTest/kotlin/me/kyd3snik/test/diff/util/TeddiSandboxTestResultCollector.kt
@@ -1,13 +1,0 @@
-package me.kyd3snik.test.diff.util
-
-import org.gradle.api.internal.tasks.testing.junit.result.TestClassResult
-import org.gradle.api.internal.tasks.testing.junit.result.TestResultSerializer
-import java.io.File
-
-class TeddiSandboxTestResultCollector(private val projectDir: File, private val moduleName: String) {
-
-    fun collectResults(): List<TestClassResult> = ArrayList<TestClassResult>().also { results ->
-        TestResultSerializer(File(projectDir, "$moduleName/build/test-results/testDebugUnitTest/binary"))
-            .read(results::add)
-    }
-}
