+++ src/main/kotlin/io/github/kartashovaa/teddi/test/resolver/TestResolver.kt
@@ -1,0 +1,10 @@
+package io.github.kartashovaa.teddi.test.resolver
+
+import org.gradle.api.file.FileCollection
+import org.gradle.api.tasks.testing.TestFilter
+
+interface TestResolver {
+
+    fun resolve(changes: FileCollection, filter: TestFilter)
+}
+
