+++ src/main/kotlin/io/github/kartashovaa/teddi/test/resolver/usage/UsageCollector.kt
@@ -1,0 +1,10 @@
+package io.github.kartashovaa.teddi.test.resolver.usage
+
+import java.io.File
+
+interface UsageCollector {
+
+    fun collect(classFile: File): Set<String>
+}
+
+
