+++ src/main/kotlin/io/github/kartashovaa/teddi/utils/CharSequenceExt.kt
@@ -1,0 +1,3 @@
+package io.github.kartashovaa.teddi.utils
+
+fun String.capitalized(): String = replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }
