+++ src/main/kotlin/io/github/kartashovaa/teddi/utils/TaskCompat.kt
@@ -1,0 +1,14 @@
+package io.github.kartashovaa.teddi.utils
+
+import org.gradle.api.Task
+import org.gradle.util.GradleVersion
+
+object TaskCompat {
+
+    fun notCompatibleWithConfigurationCache(task: Task, reason: String) {
+        if (GradleVersion.current() >= GradleVersion.version("7.4")) {
+            @Suppress("UnstableApiUsage")
+            task.notCompatibleWithConfigurationCache(reason)
+        }
+    }
+}
