--- src/main/kotlin/me/kyd3snik/test/diff/options/DiffConstraintsOptionsFacade.kt
@@ -1,12 +1,0 @@
-package me.kyd3snik.test.diff.options
-
-import org.gradle.api.tasks.options.Option
-
-interface DiffConstraintsOptionsFacade {
-
-    @Option(option = "fromBlob", description = "Start blob of changes to test")
-    fun setFromBlob(fromBlob: String)
-
-    @Option(option = "toBlob", description = "End blob of changes to test")
-    fun setToBlob(toBlob: String)
-}
