--- src/main/kotlin/me/kyd3snik/test/diff/test/resolver/FilterTestResolver.kt
@@ -1,15 +1,0 @@
-package me.kyd3snik.test.diff.test.resolver
-
-import org.gradle.api.file.FileCollection
-import org.gradle.api.tasks.testing.TestFilter
-
-class FilterTestResolver(
-    private val supportedExtensions: Set<String>,
-    private val delegate: TestResolver
-) : TestResolver {
-
-    override fun resolve(changes: FileCollection, filter: TestFilter) = delegate.resolve(
-        changes.filter { file -> file.extension in supportedExtensions },
-        filter
-    )
-}
