--- src/main/kotlin/me/kyd3snik/test/diff/test/resolver/TestResolver.kt
@@ -1,10 +1,0 @@
-package me.kyd3snik.test.diff.test.resolver
-
-import org.gradle.api.file.FileCollection
-import org.gradle.api.tasks.testing.TestFilter
-
-interface TestResolver {
-
-    fun resolve(changes: FileCollection, filter: TestFilter)
-}
-
