--- src/main/kotlin/me/kyd3snik/test/diff/test/resolver/usage/UsageCollector.kt
@@ -1,10 +1,0 @@
-package me.kyd3snik.test.diff.test.resolver.usage
-
-import java.io.File
-
-interface UsageCollector {
-
-    fun collect(classFile: File): Set<String>
-}
-
-
