--- src/main/kotlin/me/kyd3snik/test/diff/utils/CharSequenceExt.kt
@@ -1,3 +1,0 @@
-package me.kyd3snik.test.diff.utils
-
-fun String.capitalized(): String = replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }
