--- src/main/kotlin/me/kyd3snik/test/diff/utils/TaskCompat.kt
@@ -1,14 +1,0 @@
-package me.kyd3snik.test.diff.utils
-
-import org.gradle.api.Task
-import org.gradle.util.GradleVersion
-
-object TaskCompat {
-
-    fun notCompatibleWithConfigurationCache(task: Task, reason: String) {
-        if (GradleVersion.current() >= GradleVersion.version("7.4")) {
-            @Suppress("UnstableApiUsage")
-            task.notCompatibleWithConfigurationCache(reason)
-        }
-    }
-}
