+++ src/test/kotlin/io/github/kartashovaa/teddi/test/resolver/usage/KotlinCandidate.kt
@@ -1,0 +1,11 @@
+package io.github.kartashovaa.teddi.test.resolver.usage
+
+class KotlinCandidate {
+
+    val a = String.format("Hello, world")
+
+    fun test() {
+        val b = List(12) { 1 }
+        val c = Int.MAX_VALUE
+    }
+}
