--- src/test/kotlin/me/kyd3snik/test/diff/test/resolver/usage/AsmUsageCollectorTest.kt
@@ -1,17 +1,0 @@
-package me.kyd3snik.test.diff.test.resolver.usage
-
-import org.junit.Test
-import java.io.File
-
-class AsmUsageCollectorTest {
-
-    @Test
-    fun collectKtUsages() {
-        val collector = AsmUsageCollector()
-
-        val file = File("build/classes/kotlin/test", KotlinCandidate::class.java.name.replace(".", "/") + ".class")
-        println(file)
-        val usages = collector.collect(file)
-        println(usages)
-    }
-}
