--- src/test/kotlin/me/kyd3snik/test/diff/test/resolver/usage/KotlinCandidate.kt
@@ -1,11 +1,0 @@
-package me.kyd3snik.test.diff.test.resolver.usage
-
-class KotlinCandidate {
-
-    val a = String.format("Hello, world")
-
-    fun test() {
-        val b = List(12) { 1 }
-        val c = Int.MAX_VALUE
-    }
-}
